# Optimize HasTag function