@francois-berder 2021-01-28
@colincross I modified the implementation and added a benchmark for the HasTag function. This should address your comments.